agent_template_96435/requirements.txt
```
python-dotenv
langchain-nvidia-ai-endpoints
langgraph
langchain-community-tools
langchain-core
```